requirements.txt
flask
flask_httpauth
werkzeug
tensorflow
pillow
numpy
logging
gunicorn
requests